typo

+++ extra_collect.py
@@ -98,9 +98,9 @@ if experiment in ['weighting_error_lambd', 'weighting_error_neurons']:
 	n_sid = int(sys.argv[2])
 	if experiment=='weighting_error_lambd':
 		n_neurons = [int(sys.argv[3])]
-		lambdas = [int(arg) for arg in sys.argv[4:]]
+		lambdas = [float(arg) for arg in sys.argv[4:]]
 	if experiment=='weighting_error_neurons':
-		lambdas = [int(sys.argv[3])]
+		lambdas = [float(sys.argv[3])]
 		n_neurons = [int(arg) for arg in sys.argv[4:]]
 	sids = pd.read_pickle(f"data/yoo.pkl")['sid'].unique()[:n_sid]
 	label = sys.argv[-1]
